+++ changes.rst
@@ -31,6 +31,7 @@ Bug fixes:
    result in a segfault (:bugref:`901`).
 -  Fix a problem where an operator such as ``<=`` on an optional type would
    sometimes lead to an internal compiler error (:bugref:`898`).
+-  Print enum values in stack traces (:bugref:`912`).
 
 .. _v2.9.2: